feat: add `DataInterpolationsRegularizationToolsExt` extension
- Move `RegularizationSmooth` definitions to the extension
- Extension loads when RegularizationTools is loaded
